@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 import {Router} from '@angular/router';
 import { GoogleLoginProvider, SocialAuthService, SocialUser } from '@abacritt/angularx-social-login';
-import { log } from 'console';
+import { error, log } from 'console';
 import { Subscription } from 'rxjs';
 import { ResponseData } from 'src/app/shared/model/ResponseData';
 import { AuthService } from 'src/app/shared/services/auth/auth.service';
@@ -17,7 +17,7 @@ import { ToastService } from 'src/app/shared/services/toast-service/toast.servic
 export class LoginComponent implements OnInit {
   hidePassword = true;
   rememberMe = false;
-  public loginForm: FormGroup = {} as  FormGroup;
+  public loginForm: UntypedFormGroup = {} as  UntypedFormGroup;
   loginSubscription$:Subscription={} as Subscription 
   userInfo!:SocialUser
   constructor( private router: Router,
@@ -34,10 +34,10 @@ export class LoginComponent implements OnInit {
 
   /* This method is used to intialize the form  */
    intilalizeForm(){
-      this.loginForm = new FormGroup({
-        username: new FormControl('', [Validators.required, Validators.minLength(6)]),
-        password: new FormControl('', [Validators.required, Validators.minLength(6)]),
-        checkbox: new FormControl(false) 
+      this.loginForm = new UntypedFormGroup({
+        username: new UntypedFormControl('', [Validators.required, Validators.minLength(6)]),
+        password: new UntypedFormControl('', [Validators.required, Validators.minLength(6)]),
+        checkbox: new UntypedFormControl(false) 
       });
     }
 
@@ -55,10 +55,12 @@ export class LoginComponent implements OnInit {
         this.spinnerService.hideSpinner();
       console.log(response);
       let token=response.payload?.token;
-      let username=response.payload?.username
+      let username=response.payload?.username;
+      let userId=response.payload?.id
       if(token){
         localStorage.setItem("token",token)
-        localStorage.setItem("username",username)
+        localStorage.setItem("username",username);
+        localStorage.setItem("userId",userId)
         location.reload();
         this.toastService.showSuccess("Logged In SuccessFully")
         this.router.navigate(['/home'])
@@ -71,22 +73,25 @@ export class LoginComponent implements OnInit {
       
     })
 
+
+    this.authGoogleService.authState.subscribe(res=>{
+      this.userInfo=res;
+    })
   }
-  signInWithGoogle(): void {
-    console.log("hello");
+  // signInWithGoogle(): any {
+  //   console.log("hello");
     
-    this.authGoogleService?.signIn(GoogleLoginProvider.PROVIDER_ID)
-      .then((user) => {
-        // Handle successful sign-in
-        console.log(user);
-        
-      })
-      .catch((error) => {
-        // Handle error
-        console.log(error);
-        
-      });
+  //   this.authGoogleService.signIn(GoogleLoginProvider.PROVIDER_ID).then((user)=>{
+  //      this.userInfo=user;
+  //      console.log(this.userInfo);
+       
+  //   },error=>{
+  //     console.log(error);
+      
+  //   });
+  // }
+  signOut(){
+    this.authGoogleService.signOut()
   }
-  
 
 }
