@@ -1,3 +1,6 @@
 export const environment = {
-  production: true
+  production: true,
+  baseUrl:"http://18.213.137.104:8080",
+  baseAuthLoginUrl:"http://18.213.137.104:8080/api/auth/login",
+  baseAuthSignupUrl:"http://18.213.137.104:8080/api/auth/signup"
 };
